requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-allauth==65.2.0
django-ckeditor==6.7.1
django-filter==24.3
django-js-asset==2.2.0
djangorestframework==3.15.2
idna==3.10
Markdown==3.7
pillow==11.0.0
python-slugify==8.0.4
requests==2.32.3
sqlparse==0.5.2
text-unidecode==1.3
tzdata==2024.2
urllib3==2.2.3